SegmentCursor.remaining() -- cast the size from long to int; the array package [currently] only supports an int sized segment, but the underlying store supports more

+++ src/main/java/syndeticlogic/catena/array/SegmentCursor.java
@@ -55,7 +55,7 @@ public class SegmentCursor {
         if(offset == -1 || segment == null) {
             return -1;
         }
-        return segment.size() - offset;
+        return (int)segment.size() - offset;
     }
     
     public Segment segment() {
